+++ cmd/api/api.go
@@ -2,72 +2,15 @@ package api
 
 import (
 	"context"
-	. "draw-together/common"
-	"github.com/gin-gonic/gin"
-	"log"
-	"net/http"
-	"os"
-	"os/signal"
-	"syscall"
-	"time"
+	"draw-together/server/api"
 )
 
-type server interface {
-	ListenAndServe() error
-}
-
-func initServer(router *gin.Engine) *http.Server {
-	return &http.Server{
-		Addr:           Config.Server.Net.Http,
-		Handler:        router,
-		ReadTimeout:    20 * time.Second,
-		WriteTimeout:   20 * time.Second,
-		MaxHeaderBytes: 1 << 20,
-	}
-}
-
 func Start(c context.Context) {
-	Router := Routers()
-	s := initServer(Router)
+	Router := api.Routers()
+	s := api.InitServer(Router)
 	err := s.ListenAndServe()
 	if err != nil {
 		panic(err)
 	}
-	shutdown(s)
-
-}
-
-func Routers() *gin.Engine {
-	Router := gin.Default()
-
-	return Router
-}
-
-func shutdown(srv *http.Server) {
-	go func() {
-		// 开启一个goroutine启动服务
-		if err := srv.ListenAndServe(); err != nil && err != http.ErrServerClosed {
-			log.Fatalf("listen: %s\n", err)
-		}
-	}()
-
-	// 等待中断信号来优雅地关闭服务器，为关闭服务器操作设置一个5秒的超时
-	quit := make(chan os.Signal, 1) // 创建一个接收信号的通道
-	// kill 默认会发送 syscall.SIGTERM 信号
-	// kill -2 发送 syscall.SIGINT 信号，我们常用的Ctrl+C就是触发系统SIGINT信号
-	// kill -9 发送 syscall.SIGKILL 信号，但是不能被捕获，所以不需要添加它
-	// signal.Notify把收到的 syscall.SIGINT或syscall.SIGTERM 信号转发给quit
-	signal.Notify(quit, syscall.SIGINT, syscall.SIGTERM) // 此处不会阻塞
-	<-quit                                               // 阻塞在此，当接收到上述两种信号时才会往下执行
-	log.Println("Shutdown Server ...")
-	// 创建一个5秒超时的context
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	// 5秒内优雅关闭服务（将未处理完的请求处理完再关闭服务），超过5秒就超时退出
-	if err := srv.Shutdown(ctx); err != nil {
-		log.Fatal("Server Shutdown: ", err)
-	}
-
-	log.Println("Server exiting")
-
+	api.Shutdown(s)
 }
