+++ cmd/rpc/rpc_draw/draw.go
@@ -1,0 +1,1 @@
+package rpc_draw
